requirements.txt
aiogram==3.*
APScheduler==3.*
python-dotenv==1.*
tzdata==2024.*
SQLAlchemy==2.*
pytz==2024.*
